# codegen/openapi.yaml
openapi: 3.1.0
info:
  title: RPA Control-Plane API
  version: "1.0"
  description: >
    제어평면 REST API (v1). 운영 콘솔/외부 클라이언트가 호출하는 제어평면 엔드포인트의
    경로·메서드·요청/응답 요지·주요 에러코드를 고정한다. 본 문서는 api-surface.md(Control-Plane
    API Surface v1)를 1:1 변환한 D1 codegen 산출물이다.


    요지 수준 규약: components.schemas는 요지(shape) 수준이며, 상세 타입(IR/Verify/Event payload·필드
    required·examples)은 계약 ts/schema 및 codegen/types를 참조한다. 어휘 체인 abort → cancelled →
    run.cancelled, "조용한 false/unknown 금지", error-catalog 메타(httpStatus/exceptionClass)를 그대로 따른다.


    - 인증: 모든 엔드포인트 인증 필수. tenant_id는 요청 본문/쿼리가 아니라 인증 주체(토큰)에서 도출한다(api-surface §0.1).

    - 에러 본문: 모든 4xx/5xx는 ApiError(ts/error-catalog.ts). HTTP 상태코드는 ERROR_CATALOG[code].httpStatus를 그대로 사용한다(api-surface §0.2).

    - DEAD_LETTER(httpStatus 200)는 API 오류 응답이 아니다 — 상태 통지 전용 코드이므로 ApiError로 반환하지 않는다(api-surface §0.2/§4).
servers:
  - url: /v1
    description: Control-plane v1 base path.
...
security:
  - bearerAuth: []
paths:
...
  /sites/{site_profile_id}/elements:
    parameters:
      - $ref: '#/components/parameters/SiteProfileId'
    get:
      tags: [Sites]
      operationId: listSiteElements
      summary: 사이트 화면요소 저장소 목록
      description: Browser Object Repository v1. DOM-backed browser selector metadata only; desktop/vision objects are out of scope.
      parameters:
        - name: stability
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SiteElementStability'
        - name: search
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: 사이트 화면요소 저장소 목록.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteElementPage'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Sites]
      operationId: createSiteElement
      summary: 사이트 화면요소 등록
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteElementCreateRequest'
      responses:
        '201':
          description: 화면요소 등록 완료.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteElement'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /sites/{site_profile_id}/elements/{element_id}:
    parameters:
      - $ref: '#/components/parameters/SiteProfileId'
      - name: element_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Sites]
      operationId: updateSiteElement
      summary: 사이트 화면요소 수정
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteElementUpdateRequest'
      responses:
        '200':
          description: 화면요소 수정 완료.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteElement'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Sites]
      operationId: deleteSiteElement
      summary: 사이트 화면요소 삭제
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 화면요소 삭제 완료.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteElementDeleteResponse'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
  /sites/{site_profile_id}/elements/{element_id}/probe:
    parameters:
      - $ref: '#/components/parameters/SiteProfileId'
      - name: element_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Sites]
      operationId: probeSiteElement
      summary: Browser Object Repository selector probe
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteElementProbeRequest'
      responses:
        '200':
          description: Selector probe result. not_run is explicit when no live probe provider or sample URL is available.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteElementProbeResponse'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        default:
          $ref: '#/components/responses/Error'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT bearer. tenant_id 및 roles(Role[]) 클레임에서 도출(auth-rbac §3). 클라이언트 body/query의
        tenant_id는 불신. 인증 미성립(토큰 누락/서명 무효) 시 UNAUTHENTICATED(401), 인증됐으나 tenant
        누락/모호 또는 역할 권한 부족 시 AUTHZ_FORBIDDEN(403). 두 경우 모두 본문은 ApiError.
  # --- Reusable parameters ---
  parameters:
...
    SiteProfileId:
      name: site_profile_id
      in: path
      required: true
      schema: { type: string }
...
    Limit:
      name: limit
      in: query
      required: false
      description: 페이지 크기. 상한 기본값은 운영 정책(Phase 3) — 본 계약은 파라미터 형태만 고정(api-surface §0.5).
      schema:
        type: integer
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: 불투명 커서. 직전 응답 next_cursor 값을 그대로 전달(api-surface §0.5).
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        부작용 명령형 POST의 중복 제출 보호(api-surface §0.4). 서버는 (tenant_id, endpoint, key)로
        최초 결과를 보관, 동일 키 재제출 시 부작용 재실행 없이 최초 응답 반환. sink_idempotency_key(데이터평면)와는 다른 계층.
      schema:
        type: string

  # --- Reusable headers ---
...
  responses:
    Error:
      description: >
        오류 응답. 본문은 ApiError(ts/error-catalog.ts). HTTP 상태는 ERROR_CATALOG[code].httpStatus를
        그대로 사용한다. DEAD_LETTER는 오류 응답이 아니므로 본문 code로 반환되지 않는다(api-surface §0.2).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'

  # --- Schemas (요지 수준; 상세 타입은 계약 ts/schema 및 codegen/types 참조) ---
  schemas:
...
    ApiError:
      type: object
      description: 공통 에러 응답(ts/error-catalog.ts ApiError).
      required: [code, message, correlation_id]
      additionalProperties: false
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: userMessage 또는 상세(민감정보 없음).
        details:
          description: 코드별 상세 페이로드(예 IR_SCHEMA_INVALID의 reason). 구조는 D1 위임(api-surface §9).
        correlation_id:
          type: string
          format: uuid
          description: event-envelope·trace correlation_id와 동일 값(api-surface §0.2).
    ErrorCode:
      type: string
      description: >
        ts/error-catalog.ts ErrorCode(48개). DEAD_LETTER는 enum에 포함되나 상태 통지 전용으로
        ApiError 응답 code로는 반환되지 않는다(api-surface §0.2).
      enum:
        - RUN_NOT_FOUND
        - RESOURCE_NOT_FOUND
        - RUN_ALREADY_TERMINAL
        - RUN_ABORTED
        - SCENARIO_VERSION_CONFLICT
        - POLICY_VERSION_CONFLICT
        - IR_SCHEMA_INVALID
        - IR_EXPRESSION_COMPILE_ERROR
        - IR_EXPRESSION_RUNTIME
        - IR_NO_BRANCH_MATCHED
        - SITE_PROFILE_BLOCKED
        - SITE_CIRCUIT_OPEN
        - SESSION_LOCKED
        - SESSION_GENERATION_CONFLICT
        - SESSION_REGISTRATION_REQUIRED
        - CHALLENGE_UNRESOLVED
        - RATE_BUDGET_EXCEEDED
        - BROWSER_LEASE_EXPIRED
        - BROWSER_CRASH
        - CDP_DISCONNECTED
        - NAVIGATION_TIMEOUT
        - LLM_BUDGET_EXCEEDED
        - LLM_CAPABILITY_MISMATCH
        - LLM_STREAM_TIMEOUT
        - LLM_STREAM_IDLE_TIMEOUT
        - LLM_MALFORMED_OUTPUT
        - LLM_CONTENT_FILTERED
        - LLM_RATE_LIMITED
        - LLM_BACKEND_UNAVAILABLE
        - LLM_CONNECTION_FAILED
        - EXTRACT_SCHEMA_INVALID
        - VERIFY_FAILED
        - EMPTY_RESULT_NO_WITNESS
        - SECRET_ACCESS_DENIED
        - DOMAIN_POLICY_VIOLATION
        - PROMPT_INJECTION_DETECTED
        - ARTIFACT_NOT_REDACTED
        - SHELL_COMMAND_NOT_ALLOWED
        - UNAUTHENTICATED
        - AUTHZ_FORBIDDEN
        - CONNECTOR_PERMISSION_DENIED
        - CONNECTOR_INCOMPATIBLE
        - CONNECTOR_HOOK_FAILED
        - SINK_DELIVERY_FAILED
        - RAW_PERSIST_FAILED
        - CONTROL_PLANE_INTERNAL_ERROR
        - HUMAN_TASK_EXPIRED
        - APPROVAL_ALREADY_DECIDED
        - WORKITEM_CHECKOUT_CONFLICT
        - DEAD_LETTER

    # ===== State enums (ts/state-machine-types.ts — 정확히 일치) =====
...
    SiteElementType:
      type: string
      enum: [button, input, link, table, row, field, message, other]
    SiteElementStability:
      type: string
      enum: [stable, review_needed, broken]
    SiteElementSource:
      type: string
      enum: [manual, pbd, capture, imported]
    SiteElement:
      type: object
      required:
        - element_id
        - site_profile_id
        - element_key
        - label
        - selector
        - element_type
        - stability
        - source
        - sample_url
        - notes
        - usage_count
        - last_verified_at
        - updated_by
        - created_at
        - updated_at
      properties:
        element_id:
          type: string
          format: uuid
        site_profile_id:
          type: string
          format: uuid
        element_key:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9_]{1,63}$'
        label:
          type: string
        selector:
          type: string
        element_type:
          $ref: '#/components/schemas/SiteElementType'
        stability:
          $ref: '#/components/schemas/SiteElementStability'
        source:
          $ref: '#/components/schemas/SiteElementSource'
        sample_url:
          type: [string, "null"]
        notes:
          type: [string, "null"]
        usage_count:
          type: integer
          minimum: 0
        last_verified_at:
          type: [string, "null"]
          format: date-time
          description: Last successful or explicit selector probe timestamp; managed by the probe endpoint, not by generic PATCH.
        updated_by:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      additionalProperties: false
    SiteElementPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SiteElement'
        next_cursor:
          type: [string, "null"]
      additionalProperties: false
...
    SiteElementCreateRequest:
      type: object
      required: [element_key, label, selector]
      properties:
        element_key:
          type: string
          pattern: '^[A-Za-z][A-Za-z0-9_]{1,63}$'
        label:
          type: string
          minLength: 1
        selector:
          type: string
          minLength: 1
        element_type:
          $ref: '#/components/schemas/SiteElementType'
        stability:
          $ref: '#/components/schemas/SiteElementStability'
        source:
          $ref: '#/components/schemas/SiteElementSource'
        sample_url:
          type: string
        notes:
          type: string
      additionalProperties: false
    SiteElementUpdateRequest:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        selector:
          type: string
          minLength: 1
        element_type:
          $ref: '#/components/schemas/SiteElementType'
        stability:
          $ref: '#/components/schemas/SiteElementStability'
        sample_url:
          type: [string, "null"]
        notes:
          type: [string, "null"]
      additionalProperties: false
    SiteElementProbeStatus:
      type: string
      enum: [matched, not_found, invalid_selector, failed, not_run]
    SiteElementProbeRequest:
      type: object
      properties:
        sample_url:
          type: string
      additionalProperties: false
    SiteElementProbeResponse:
      type: object
      required: [element_id, site_profile_id, selector, sample_url, probe_status, match_count, reason_code, checked_at, element]
      properties:
        element_id:
          type: string
          format: uuid
        site_profile_id:
          type: string
          format: uuid
        selector:
          type: string
        sample_url:
          type: [string, "null"]
        probe_status:
          $ref: '#/components/schemas/SiteElementProbeStatus'
        match_count:
          type: [integer, "null"]
          minimum: 0
        reason_code:
          type: [string, "null"]
        checked_at:
          type: string
          format: date-time
        element:
          $ref: '#/components/schemas/SiteElement'
      additionalProperties: false
...
    SiteElementDeleteResponse:
      type: object
      required: [element_id, deleted]
      properties:
        element_id:
          type: string
          format: uuid
        deleted:
          type: boolean
      additionalProperties: false
...